backend/requirements.txt
Flask
mysql-connector-python
google-cloud-speech
google-cloud-aiplatform